requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiocache==0.12.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.11.0
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
billiard==4.2.0
boto3==1.34.131
botocore==1.34.131
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.1
fastapi-limiter==0.1.6
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.4
iniconfig==2.0.0
jiter==0.5.0
jmespath==1.0.1
kombu==5.4.0
limits==3.13.0
multidict==6.0.5
numpy==2.1.0
openai==1.42.0
packaging==24.1
passlib==1.7.4
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.8
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
